+++ packages/composites/crypto-volatility-index/src/adapter.ts
@@ -1,5 +1,5 @@
-import { Execute } from '@chainlink/types'
 import { Requester, Validator } from '@chainlink/ea-bootstrap'
+import { AdapterRequest, Execute } from '@chainlink/types'
 import { calculate } from './cryptoVolatilityIndex'
 
 const customParams = {
@@ -7,19 +7,21 @@ const customParams = {
   multiply: false,
   heartbeatMinutes: false,
   isAdaptive: false,
+  cryptoCurrencies: false,
+  deviationThreshold: false,
+  lambdaMin: false,
+  lambdaK: false,
 }
 
-export const execute: Execute = async (input) => {
+export const execute: Execute = async (input: AdapterRequest) => {
   const validator = new Validator(input, customParams)
   if (validator.error) throw validator.error
 
   const jobRunID = validator.validated.id
 
   const result = await calculate(validator.validated, input.data)
-  return Requester.success(jobRunID, {
-    data: { result },
-    status: 200,
-  })
+  const response = { data: { result }, status: 200 }
+  return Requester.success(jobRunID, response)
 }
 
 export default execute